README.md: Update screenshot
GitHub updated the design, shrinking the space available for user
content in Readme documents. This caused the screenshot to be
shrinked and blurry, due to browsers squeezing the image into a space
smaller than it.

Re-layout the two panes to fit to the new design width (882 pixels) to
fix this blurry rendering.