Add option to skip package file checksum checks on save
Until now, when updating the configuration to reflect the state of the
system with aconfmgr save, the "Searching for modified files..." phase
verifies that the checksum of each file associated with an installed
package matches the checksum stored in pacman's database.

While this perfectly ensures that changes to the package's files are
always detected, this requires reading the entire contents of all files
associated to installed packages on the system, which generally amounts
to several gigabytes. This makes aconfmgr's execution slow, even on
systems with a capable CPU and a SSD, which makes routine updates and
checks bothersome.

However, in general such checksum checks are unnecessary, since
modifying a file associated with an installed package by usual means
will modify its properties, most importantly its modification time and
size, so that changes can still be detected without needing to execute
a costly checksum check. Accordingly, other tools such as rsync also
avoid a full checksum check and generally rely on changes on
modification time and file size to detect changes efficiently.

This commit adds a --skip-checksum command line switch, which applies
to aconfmgr save, which allows skipping the checksum checks on the
installed package's files. This switch is not enabled by default, so
the default behavior of checking each package file's checksum is kept.

Note that it's technically possible that a file still has all the same
properties, including modification time and size, while actually having
different file contents. In such as case, using this option will
erroneously not detect this file as modified. However, generally this
problem only manifests under exceptional circumstances, such as badly
behaved programs, intentional manipulation or filesystem corruption.

Fixes #51.