src/sudo: Don't wait for cat from terminal to finish
Read from terminal stdin is going to block forever and will "never"
EOF. We still need to pass stdin data (i.e. anything the user types as
long as the command is running) for things like pacman confirmations,
but we do have to kill the cat once the program exits.