Add tests for the --skip-checksums option
This commit adds unit tests for the recently introduced feature to skip
package file checksum checks on save.

Two tests exercise the usual cases where updates to an installed
package's file are detected through changes to size or modification
time respectively, while a third test exercises the exceptional case
(as explained in the commit introducing the feature) where a change
to an installed package's file is not detected due to the file having
all the same properties (size, modification time, owner, group, etc.)
as the package's original file.

As part of the implementation of the tests, a trivial bug in
lib-funcs-integ.bash:TestAddFSObj has been fixed, which until now was
harmless, but impacted the tests introduced in this commit.
Similarly, an existing test which was not following the typical naming
convention has been renamed without changes, in order to follow it.

See issue #51.