Speed up lost files search for large number of ignore paths
When updating the system configuration with aconfmgr save, if a large
number of ignored path patterns is used in the configuration (through
the IgnorePath helper function), the "Searching for lost files..."
phase can get considerably slow, even taking up most of the execution
time if I/O is fast (such as when using a fast SSD disk).

This happens because ignored path patterns are finally given as
-wholename options to GNU find (the find implementation used in Arch),
and as an implementation detail of this tool, the search time seems to
grow linearly with the number of search patterns.

However, also as an implementation detail of this tool, the -regex
option scales much better. Therefore, passing the patterns using this
option is substantially faster. However, of course, since IgnorePath
accepts shell patterns, we need to find a way to convert those to
regular expressions.

A general shell pattern to regular expression conversion approach,
while technically possible, is very complex, since there are many
small details and edge cases where shell patterns and regular
expressions differ. However, in practice, the majority of IgnorePath
shell patterns we expect to find are 'simple' ones, made of simple,
literal ASCII characters (e.g. alphanumeric, underscore, hyphen, slash)
plus the asterisk wildcard.

This commit adds some logic to detect which IgnorePath shell patterns
are 'simple' (as explained previously), and converts them to an
equivalent regular expression which is then provided to GNU find. This
notably speeds up the execution due to the workings of GNU find.
The non-'simple' patterns are still passed as shell patterns to find
like they were before. The net result is that the optimization is
applied transparently and should have no visible effect on the results.