WIP: describe concepts first.
The readme has a great, concise introduction, but the Usage section
jumps pretty hard into "first run" specifics without explaining
some needed background information.   I feel like there should be
some conceptual descriptions before getting into the specifics of
the commands to run.  Defining some concepts and terms early on
means the rest of the docs can be less verbose.

This is an initial attempt at doing so, and while doing it,
it made sense to move the modus operandi into it.
I'm filing this PR as a starting point for a conversation around it,
it should not be merged as is.

this builds up understanding more logically
(but people can still jump ahead if they choose to using the TOC).

Looking forward, we should probably split up some stuff into separate
pages.  E.g:

* keep the README lean / headline stuff.
* separate page: tutorial / first run
* separate page for concepts deep dive
* separate howto's (e.g. ignoring files, scripting)
* separate page for comparison to other systems