Add support for colons in package descriptions
With the previous implementation, `aconfmgr save` would generate config
files with a truncated `AddPackage` comment if the package description
contains a colon character. For example:

    $ pacman -S potrace
    $ aconfmgr save
    $ cat 99-unsorted.sh
    AddPackage potrace # Utility for tracing a bitmap (input

This is because of the `cut` incantation we had:

    $ LC_ALL=C pacman --query --info "potrace" | grep '^Description' | cut -d ':' -f 2
     Utility for tracing a bitmap (input

But the description of potrace is actually longer than that:

    $ LC_ALL=C pacman --query --info "potrace" | grep '^Description'
    Description     : Utility for tracing a bitmap (input: PBM,PGM,PPM,BMP; output: EPS,PS,PDF,SVG,DXF,PGM,Gimppath,XFig)

Fortunately, `cut` has a pretty simple tweak to get everything from the
Nth field to the end of the line. From `man cut`:

    > N-     from N'th byte, character or field, to end of line

And this seems to work great!

    $ LC_ALL=C pacman --query --info "potrace" | grep '^Description' | cut -d ':' -f 2-
     Utility for tracing a bitmap (input: PBM,PGM,PPM,BMP; output: EPS,PS,PDF,SVG,DXF,PGM,Gimppath,XFig)