Revert "common.bash: Automate pacman prompts more aggressively"
This reverts commit cd1c4b0f46c23ddf763f37f9d6c930f272690d9f.

Not a universal solution as pacman asks for not just yes-or-no
questions, see e.g.
https://github.com/CyberShadow/aconfmgr/issues/143.

Let's park this feature until we have a compelling, test-case-backed
use case for it.  For now, users will need to occasionally run
aconfmgr in interactive mode to confirm prompts that pacman deems
too dangerous for --noconfirm.